requirements.txt
django == 3.1.2
django-admin-interface
django-crispy-forms
pillow
django-rest-framework
django-paypal
paypal-checkout-serversdk
pymysql
mysqlclient
django-jazzmin

##  pip install -r requirements.txt